@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class DrivesCandidate < ApplicationRecord
   belongs_to :drive
   belongs_to :candidate
@@ -7,11 +9,10 @@ class DrivesCandidate < ApplicationRecord
     self.email_sent_at = Time.now.utc
     save!
   end
-  
+
   private
-  
+
   def generate_token
     SecureRandom.hex(20)
   end
 end
-
