@@ -1,12 +1,15 @@
+# frozen_string_literal: true
+
 class User < ApplicationRecord
-  # Include default devise modules. Others available are:
-  # :confirmable, :lockable, :timeoutable, :trackable and :omniauthable
-  devise :database_authenticatable, 
-         :recoverable, :rememberable, :validatable
+  include DeviseTokenAuth::Concerns::User
+
+  # Include default devise modules.
+  devise :database_authenticatable
+  before_save -> { skip_confirmation! }
+
   belongs_to :organization
   belongs_to :role
   has_many :problems
   has_many :drives
   has_many :test_cases
 end
-
