@@ -1,18 +1,20 @@
-require "http"
+# frozen_string_literal: true
+
+require 'http'
 
 class JudgeZeroApi
-  BASE_URI = 'http://localhost:3000'
+  BASE_URI = 'http://65.1.201.245'
 
-  def initialize(params={})
+  def initialize(params = {}, headers = {})
     @params = params
+    @headers = headers
   end
 
   def get(path)
     response = HTTP.get("#{BASE_URI}#{path}")
-    response = JSON.parse response
   end
 
   def post(path)
-    HTTP.post("#{BASE_URI}#{path}", params: @params)
+    HTTP.post("#{BASE_URI}#{path}", headers: @headers, body: @params.to_json)
   end
 end
