@@ -1,9 +1,11 @@
+# frozen_string_literal: true
+
 FactoryBot.define do
   factory :candidate, class: Candidate do
+    first_name { Faker::Name.name }
+    last_name { Faker::Name.name  }
     email { Faker::Internet.email }
-    first_name { 'Samruddhi' }
-    last_name { 'Deshpande' }
-    is_profile_complete { 'Yes' }
-    drive_id { 9 }
+    mobile_number { Faker::PhoneNumber }
+    is_profile_complete { true }
   end
 end
