@@ -1,8 +1,11 @@
+# frozen_string_literal: true
+
 FactoryBot.define do
   factory :drive, class: Drive do
-    name { 'drive1' }
-    description { 'Drive details' }
-    created_at { Time.now }
-    updated_at { Time.now }
+    name { 'TestingDrive' }
+    description { 'Drive created for RSpec Testing' }
+    start_time { Time.now.localtime }
+    end_time { Time.now.localtime + 3.hours }
+    duration { end_time - start_time }
   end
 end
